put all the output files in build folder
