+++ index.html
@@ -9,8 +9,8 @@
   <script src="node_modules/vega-lite/vega-lite.js"></script>
   <script src="node_modules/vega-embed/vega-embed.js" charset="utf-8"></script>
 
-  <script src="index.js"></script>
-  <link rel="stylesheet" type="text/css" href="vg-tooltip.css">
+  <script src="build/index.js"></script>
+  <link rel="stylesheet" type="text/css" href="build/vg-tooltip.css">
 
   <!-- Example Page -->
   <script src="examples.js"></script>
